requirements.txt
mysql_connector_repackaged
pandas
plotly
SQLAlchemy
streamlit




